+++ magik-squid/src/main/java/nl/ramsolutions/sw/magik/analysis/typing/types/Parameter.java
@@ -91,8 +91,19 @@ public class Parameter {
         return this.modifier;
     }
 
-    public boolean is(Modifier isModifier) {
-        return this.modifier == isModifier;
+    /**
+     * Test if parameter has modifier.
+     * @param isModifiers Modifiers to test.
+     * @return True if parameter has modifier.
+     */
+    public boolean is(Modifier... isModifiers) {
+        for (final Modifier isModifier : isModifiers) {
+            if (this.modifier == isModifier) {
+                return true;
+            }
+        }
+
+        return false;
     }
 
     public void setType(TypeString type) {